Emotion_Engine/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asyncio==4.0.0
certifi==2025.11.12
cffi==2.0.0
cryptography==46.0.3
deprecation==2.1.0
filelock==3.20.0
fsspec==2025.12.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.6
numpy==2.3.5
packaging==25.0
pillow==12.0.0
postgrest==2.25.0
propcache==0.4.1
protobuf==6.33.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
realtime==2.25.0
rich==14.2.0
setuptools==80.9.0
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
websockets==15.0.1
yarl==1.22.0
